pyproject.toml
[tool.poetry]
name = "sendbox-auth"
version = "0.1.0"
description = ""
authors = ["Anton Petrov <marblemonolit@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.1.1"
Flask-Migrate = "^3.1.0"
Flask-SQLAlchemy = "^2.5.1"
Flask-JWT-Extended = "^4.3.1"
Werkzeug = "^2.1.1"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
ipython = "^8.2.0"
ipdb = "^0.13.9"
pytest-cov = "^3.0.0"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
Faker = "^13.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
